Move cross platform _pipe wrapper to utils.h
