+++ triangulation.h
@@ -20,7 +20,6 @@ public:
     triangulation(){};
     triangulation(vector<v2f> Points, vector<v2f> Region);
     void create();
-    
     void refine(double tol);
     triangle StartingTriangle;
     vector<v2f> Points;
